# backend/docs/swagger.yaml
basePath: /
definitions:
  github_com_Innocent9712_much-to-do_Server_MuchToDo_internal_models.ChangePasswordDTO:
    properties:
      newPassword:
        minLength: 8
        type: string
      oldPassword:
        type: string
    required:
    - newPassword
    - oldPassword
    type: object
...
  github_com_Innocent9712_much-to-do_Server_MuchToDo_internal_models.LoginUserDTO:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  github_com_Innocent9712_much-to-do_Server_MuchToDo_internal_models.PublicUser:
    properties:
      firstName:
        type: string
      id:
        type: string
      lastName:
        type: string
      username:
        type: string
    type: object
  github_com_Innocent9712_much-to-do_Server_MuchToDo_internal_models.RegisterUserDTO:
    properties:
      firstName:
        type: string
      lastName:
        type: string
      password:
        minLength: 6
        type: string
      username:
        minLength: 3
        type: string
    required:
    - firstName
    - lastName
    - password
    - username
    type: object
...
  github_com_Innocent9712_much-to-do_Server_MuchToDo_internal_models.UpdateUserDTO:
    properties:
      firstName:
        type: string
      lastName:
        type: string
      username:
        type: string
    type: object
info:
  contact:
    email: innocent@altschoolafrica.com
    name: API Support - Innocent
    url: https://github.com/Innocent9712
  description: This is an API for MuchToDo application with user authentication.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: MuchToDo API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: |-
        Logs in a user with username and password, returning a session token.
        The token is returned in the response body and as an httpOnly cookie.
      parameters:
      - description: User Login Credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/github_com_Innocent9712_much-to-do_Server_MuchToDo_internal_models.LoginUserDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Returns a success message, the JWT token, and user details
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Invalid username or password
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Log in a user
      tags:
      - auth
...
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account with the provided details
      parameters:
      - description: User Registration Info
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/github_com_Innocent9712_much-to-do_Server_MuchToDo_internal_models.RegisterUserDTO'
      produces:
      - application/json
      responses:
        "201":
          description: '{''message'': ''User registered successfully''}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Username is already taken
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - auth
  /auth/username-check/{username}:
    get:
      consumes:
      - application/json
      description: Checks the database and cache to see if a username is already in
        use
      parameters:
      - description: Username to check
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Returns true if the username is available
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Username is too short
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Database error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Check if a username is available
      tags:
      - auth
...
  /users/me:
    delete:
      description: Permanently deletes the authenticated user's account and all their
        associated data (e.g., todos)
      produces:
      - application/json
      responses:
        "200":
          description: '{''message'': ''Account deleted successfully''}'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete current user's account
      tags:
      - users
    get:
      description: Retrieves the profile of the user corresponding to the provided
        JWT
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Innocent9712_much-to-do_Server_MuchToDo_internal_models.PublicUser'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get current user's profile
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Updates the first name, last name, and/or username of the authenticated
        user
      parameters:
      - description: User Update Info
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/github_com_Innocent9712_much-to-do_Server_MuchToDo_internal_models.UpdateUserDTO'
      produces:
      - application/json
      responses:
        "200":
          description: '{''message'': ''Profile updated successfully''}'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Username is already taken
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update current user's profile
      tags:
      - users
  /users/me/password:
    put:
      consumes:
      - application/json
      description: Allows an authenticated user to change their password
      parameters:
      - description: Old and New Passwords
        in: body
        name: passwords
        required: true
        schema:
          $ref: '#/definitions/github_com_Innocent9712_much-to-do_Server_MuchToDo_internal_models.ChangePasswordDTO'
      produces:
      - application/json
      responses:
        "200":
          description: '{''message'': ''Password changed successfully''}'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input or validation error
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized or incorrect old password
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Change current user's password
      tags:
      - users
securityDefinitions:
  ApiKeyAuth:
    description: '"Type ''Bearer'' followed by a space and a JWT token."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"